Extract Item.Merge and Item.Split
